+++ ncollide_geometry/query/contacts_internal/contact.rs
@@ -1,8 +1,10 @@
 use std::mem;
 use math::Point;
 
+use na::{Real, Unit};
+
 /// Geometric description of a contact.
-#[derive(Debug, PartialEq, Clone, RustcEncodable, RustcDecodable)]
+#[derive(Debug, PartialEq, Clone)]
 pub struct Contact<P: Point> {
     /// Position of the contact on the first object. The position is expressed in world space.
     pub world1: P,
@@ -11,7 +13,7 @@ pub struct Contact<P: Point> {
     pub world2: P,
 
     /// Contact normal
-    pub normal: P::Vector,
+    pub normal: Unit<P::Vector>,
 
     /// Penetration depth
     pub depth: P::Real,
@@ -20,7 +22,7 @@ pub struct Contact<P: Point> {
 impl<P: Point> Contact<P> {
     /// Creates a new contact.
     #[inline]
-    pub fn new(world1: P, world2: P, normal: P::Vector, depth: P::Real) -> Contact<P> {
+    pub fn new(world1: P, world2: P, normal: Unit<P::Vector>, depth: P::Real) -> Contact<P> {
         Contact {
             world1: world1,
             world2: world2,
@@ -36,5 +38,24 @@ impl<P: Point> Contact<P> {
     pub fn flip(&mut self) {
         mem::swap(&mut self.world1, &mut self.world2);
         self.normal = -self.normal;
+    }
+}
+
+/// The prediction parameters for contact determination.n
+pub struct ContactPrediction<N: Real> {
+    /// The linear prediction.
+    pub linear: N,
+    /// The angular regularization for the first solid.
+    pub angular1: N,
+    /// The angular regularization for the second solid.
+    pub angular2: N,
+}
+
+impl<N: Real> ContactPrediction<N> {
+    /// Initialize prediction parameters.
+    pub fn new(linear: N, angular1: N, angular2: N) -> Self {
+        ContactPrediction {
+            linear, angular1, angular2
+        }
     }
 }